feat: require world selection for DND
